Refactoring the routes for flask completed
