+++ backend-flask/lib/rollbar.py
@@ -2,13 +2,14 @@
 from time import strftime
 from flask import got_request_exception
 from flask import current_app as app
+from flask import request, g
 import os
 import rollbar
 import rollbar.contrib.flask
 
 
 # Rollbar init code
-def init_rollbar():
+def init_rollbar(app):
   rollbar_access_token = os.getenv('ROLLBAR_ACCESS_TOKEN')
   rollbar.init(
     # access token
